backend/requirements.txt
Django==5.0.7
djangorestframework==3.16.0
django-cors-headers==4.7.0
psycopg2-binary==2.9.10
gunicorn==23.0.0
Pillow==10.2.0
